UI: Mark Playing Songin in Track List
This is a feature requested in issue #48
Highlight which track is playing

The solution is the one suggested in the comments
on the issue.

1. Daft Punk - Derezze
2. <<Daft Punk - The Son of Flynn>>
3. Daft Punk - TRON Legacy (End Titles)

cui.TrackPlaying() now calls updateTracksView()
updateTracksView now saves cursor and origin state and
restore the state after redrawing the list.

Signed-Off-By Zach Brown <brownzach125@gmail.com>