Feature: Create event newTrackLoaded
The newTrackLoaded event is emitted by the spotify_actions play
function, whenever it loads a new track. The channel is passed the
duration of the new track. The mainloop listens for this new event.

I also made some small changes to how the other channels are created so
they won't block, and data won't be lost just because no one is
currently listening.